http2,perf_hooks: use PERFORMANCE_RELATIVE macro for PerformanceEntry
This removes the calculation of a relative timestamp from the method
`startTimeNano()` of the `PerformanceEntry` class and changes all call-
sites to use the newly introduced PERFORMANCE_RELATIVE() macro instead.

This is done in preparation to support an additional `markOptions`
object as an argument to the `performance.mark()` method in which a user
can provide their own timestamp as `startTime`.

We tried to calculate the relative startTime in nanoseconds inside the
`v8::performance::Mark()` method instead which lead to complications and
an integer overflow of a `uint64` type.