perf_hooks: performance.mark accepts optional markOptions
With this change `performance.mark()` accepts an optinal `markOptions`
object, allowing to provide a custom `startTime` value that should be
used as `startTime` of the `PerformanceEntry`.

TODO: We still need to do some type coercion and error checking here,
also the error message needs to be adapted to the correct wording and
stored in the appropriate place.